+++ config/query-builder.php
@@ -29,4 +29,10 @@ return [
      */
     'count_suffix' => 'Count',
 
+    /*
+     * By default the package will throw an `InvalidFilterQuery` exception when a filter in the
+     * URL is not allowed in the `allowedFilters()` method.
+     */
+    'disable_invalid_filter_query_exception' => false,
+
 ];